Othello Project
This is my Othello Project where a user can create an Othello game board and play the game. The user is initially asked to input the rows and colums for their board. They are then asked which player gets to go first, player "B" or player "W". After selecting the first player, the game mode is then determined. If entered "<", the player with the least amount of pieces on the board will win. On the other hand, if the user enters ">", the player with the most pieces will win. You are finally asked to create the board by typing out the string equivalent of the board (etc: ". . . ." ". B W ." ". W B ." ". . . ."). The game will then begin.